Refactor: move setupContainer to container.go
