+++ node/container.go
@@ -15,9 +15,12 @@ package node
 
 import (
 	"fmt"
+	"math/rand"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strconv"
+	"strings"
 	"time"
 
 	log "github.com/Sirupsen/logrus"
@@ -26,7 +29,9 @@ import (
 	"github.com/control-center/serviced/commons/iptables"
 	"github.com/control-center/serviced/dfs/registry"
 	"github.com/control-center/serviced/domain/service"
+	"github.com/control-center/serviced/domain/user"
 	"github.com/control-center/serviced/rpc/master"
+	"github.com/control-center/serviced/utils"
 	"github.com/control-center/serviced/zzk"
 	zkservice "github.com/control-center/serviced/zzk/service"
 	dockerclient "github.com/fsouza/go-dockerclient"
@@ -424,4 +429,277 @@ func dockerLogsToFile(containerid string, numlines int) {
 		"file":  fname,
 		"lines": numlines,
 	}).Infof("Container log dumped to file")
+}
+
+// setupContainer creates and populates two structures, a docker client Config and a docker client HostConfig structure
+// that are used to create and start a container respectively. The information used to populate the structures is pulled from
+// the service, serviceState, and conn values that are passed into setupContainer.
+func (a *HostAgent) setupContainer(masterClient master.ClientInterface, svc *service.Service, instanceID int, imageName string) (*dockerclient.Config, *dockerclient.HostConfig, error) {
+	logger := plog.WithFields(log.Fields{
+		"serviceName": svc.Name,
+		"serviceID": svc.ID,
+		"instanceID": instanceID,
+		"imageID": imageName,
+	})
+	var err error
+	svc, err = masterClient.GetEvaluatedService(svc.ID, instanceID)
+	if err != nil {
+		logger.WithError(err).Error("Failed to get service")
+		return nil, nil, err
+	}
+	// Update the service with the complete image name
+	svc.ImageID = imageName
+
+	cfg := &dockerclient.Config{}
+	hcfg := &dockerclient.HostConfig{}
+
+	//get this service's tenantId for volume mapping
+	var tenantID string
+	tenantID, err = masterClient.GetTenantID(svc.ID)
+	if err != nil {
+		logger.WithError(err).Error("Failed to get the tenant ID for the service")
+		return nil, nil, err
+	} else if len(tenantID) == 0 && len(svc.Volumes) > 0 {
+		// FIXME: find a better way of handling this error condition
+		logger.Fatal("Could not get tenant ID and need to mount a volume")
+	}
+
+	// get the system user
+	var systemUser user.User
+	systemUser, err = masterClient.GetSystemUser()
+	if err != nil {
+		logger.WithError(err).Error("Failed to get the system user account")
+		return nil, nil, err
+	}
+	logger.WithField("systemUser", systemUser.Name).Debug("Got system user")
+
+	cfg.User = "root"
+	cfg.WorkingDir = "/"
+	cfg.Image = svc.ImageID
+
+	// get the endpoints
+	cfg.ExposedPorts = make(map[dockerclient.Port]struct{})
+	hcfg.PortBindings = make(map[dockerclient.Port][]dockerclient.PortBinding)
+
+	if svc.Endpoints != nil {
+		logger.WithField("endpoints", svc.Endpoints).Debug("Endpoints for service")
+		for _, endpoint := range svc.Endpoints {
+			if endpoint.Purpose == "export" { // only expose remote endpoints
+				var p string
+				switch endpoint.Protocol {
+				case commons.UDP:
+					p = fmt.Sprintf("%d/%s", endpoint.PortNumber, "udp")
+				default:
+					p = fmt.Sprintf("%d/%s", endpoint.PortNumber, "tcp")
+				}
+				cfg.ExposedPorts[dockerclient.Port(p)] = struct{}{}
+				hcfg.PortBindings[dockerclient.Port(p)] = append(hcfg.PortBindings[dockerclient.Port(p)], dockerclient.PortBinding{})
+			}
+		}
+	}
+
+	bindsMap := make(map[string]string) // map to prevent duplicate path assignments. Use to populate hcfg.Binds later.
+
+	// iterate svc.Volumes - create bindings for non-dfs volumes
+	for _, volume := range svc.Volumes {
+		if volume.Type != "" && volume.Type != "dfs" {
+			continue
+		}
+
+		resourcePath, err := a.setupVolume(tenantID, svc, volume)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		resourcePath = strings.TrimSpace(resourcePath)
+		containerPath := strings.TrimSpace(volume.ContainerPath)
+		bindsMap[containerPath] = resourcePath
+	}
+
+	// mount serviced path
+	dir, _, err := ExecPath()
+	if err != nil {
+		logger.WithError(err).Error("Error getting the path to the serviced executable")
+		return nil, nil, err
+	}
+
+	dir, binary := filepath.Split(a.controllerBinary)
+	resourcePath := strings.TrimSpace(dir)
+	containerPath := strings.TrimSpace("/serviced")
+	bindsMap[containerPath] = resourcePath
+
+	// bind mount everything we need for filebeat
+	if len(svc.LogConfigs) != 0 {
+		const LOGSTASH_CONTAINER_DIRECTORY = "/usr/local/serviced/resources/logstash"
+		logstashPath := utils.ResourcesDir() + "/logstash"
+		resourcePath := strings.TrimSpace(logstashPath)
+		containerPath := strings.TrimSpace(LOGSTASH_CONTAINER_DIRECTORY)
+		bindsMap[containerPath] = resourcePath
+		logger.WithField("bindmount", fmt.Sprintf("%s:%s", resourcePath, containerPath)).Debug("Added logstash bindmount")
+	}
+
+	// specify temporary volume paths for docker to create
+	tmpVolumes := []string{"/tmp"}
+	for _, volume := range svc.Volumes {
+		if volume.Type == "tmp" {
+			tmpVolumes = append(tmpVolumes, volume.ContainerPath)
+		}
+	}
+
+	// add arguments to mount requested directory (if requested)
+	logger.Debug("Checking service's mount options")
+	for _, bindMountString := range a.mount {
+		logger.WithField("bindmount", bindMountString).Debug("Checking bindmount string")
+		splitMount := strings.Split(bindMountString, ",")
+		numMountArgs := len(splitMount)
+
+		if numMountArgs == 2 || numMountArgs == 3 {
+
+			requestedImage := splitMount[0]
+			hostPath := splitMount[1]
+			// assume the container path is going to be the same as the host path
+			containerPath := hostPath
+
+			// if the container path is provided, use it
+			if numMountArgs > 2 {
+				containerPath = splitMount[2]
+			}
+			logger.WithFields(log.Fields{
+				"requestedImage": requestedImage,
+				"hostPath": hostPath,
+				"containerPath": containerPath,
+			}).Debug("Parsed out bind mount information")
+
+			// insert tenantId into requestedImage - see facade.DeployService
+			matchedRequestedImage := false
+			if requestedImage == "*" {
+				matchedRequestedImage = true
+			} else {
+				imageID, err := commons.ParseImageID(requestedImage)
+				if err != nil {
+					logger.WithError(err).
+					WithField("requestedImageID", requestedImage).
+					Error("Unable to parse requested ImageID")
+					continue
+				}
+				svcImageID, err := commons.ParseImageID(svc.ImageID)
+				if err != nil {
+					logger.WithError(err).Error("Unable to parse service imageID")
+					continue
+				}
+				matchedRequestedImage = (imageID.Repo == svcImageID.Repo)
+			}
+
+			logger.WithFields(log.Fields{
+				"matchedRequestedImage": matchedRequestedImage,
+			}).Debug("Finished evaluation for matchedRequestedImage")
+
+			if matchedRequestedImage {
+				hostPath = strings.TrimSpace(hostPath)
+				containerPath = strings.TrimSpace(containerPath)
+				bindsMap[containerPath] = hostPath
+			}
+		} else {
+			logger.WithField("bindMount", bindMountString).
+			Warn("Could not bind mount the requested mount point")
+		}
+	}
+
+	// transfer bindsMap to hcfg.Binds
+	hcfg.Binds = []string{}
+	for containerPath, hostPath := range bindsMap {
+		binding := fmt.Sprintf("%s:%s", hostPath, containerPath)
+		hcfg.Binds = append(hcfg.Binds, binding)
+	}
+
+	// Get host IP
+	ips, err := utils.GetIPv4Addresses()
+	if err != nil {
+		logger.WithError(err).Error("Unable to get host IP addresses")
+		return nil, nil, err
+	}
+
+	// XXX: Hopefully temp fix for CC-1384 & CC-1631 (docker/docker issue 14203).
+	count := rand.Intn(128)
+	fix := ""
+	for i := 0; i < count; i++ {
+		fix += "."
+	}
+	// End temp fix part 1. See immediately below for part 2.
+
+	// add arguments for environment variables
+	cfg.Env = append(svc.Environment,
+		fmt.Sprintf("CONTROLPLANE_SYSTEM_USER=%s", systemUser.Name),
+		fmt.Sprintf("CONTROLPLANE_SYSTEM_PASSWORD=%s", systemUser.Password),
+		fmt.Sprintf("CONTROLPLANE_HOST_IPS='%s'", strings.Join(ips, " ")),
+		fmt.Sprintf("SERVICED_VIRTUAL_ADDRESS_SUBNET=%s", a.virtualAddressSubnet),
+		fmt.Sprintf("SERVICED_IS_SERVICE_SHELL=false"),
+		fmt.Sprintf("SERVICED_NOREGISTRY=%s", os.Getenv("SERVICED_NOREGISTRY")),
+		fmt.Sprintf("SERVICED_SERVICE_IMAGE=%s", svc.ImageID),
+		fmt.Sprintf("SERVICED_MAX_RPC_CLIENTS=1"),
+		fmt.Sprintf("SERVICED_RPC_PORT=%s", a.rpcport),
+		fmt.Sprintf("SERVICED_LOG_ADDRESS=%s", a.logstashURL),
+		//The SERVICED_UI_PORT environment variable is deprecated and services should always use port 443 to contact serviced from inside a container
+		"SERVICED_UI_PORT=443",
+		fmt.Sprintf("SERVICED_MASTER_IP=%s", strings.Split(a.master, ":")[0]),
+		fmt.Sprintf("TZ=%s", os.Getenv("TZ")),
+		// XXX: Hopefully temp fix for CC-1384 & CC-1631 (docker/docker issue 14203).
+		fmt.Sprintf("DOCKER_14203_FIX='%s'", fix),
+		// End temp fix part 2. See immediately above for part 1.
+	)
+
+	// add dns values to setup
+	for _, addr := range a.dockerDNS {
+		_addr := strings.TrimSpace(addr)
+		if len(_addr) > 0 {
+			cfg.DNS = append(cfg.DNS, addr)
+		}
+	}
+
+	// Add hostname if set
+	if svc.Hostname != "" {
+		cfg.Hostname = svc.Hostname
+	}
+
+	cmd := []string{filepath.Join("/serviced", binary)}
+
+	// Flag TLS for the mux if it's disabled
+	if !a.useTLS {
+		cmd = append(cmd, "--mux-disable-tls")
+	}
+
+	cfg.Cmd = append(cmd,
+		svc.ID,
+		strconv.Itoa(instanceID),
+		svc.Startup)
+
+	if svc.Privileged {
+		hcfg.Privileged = true
+	}
+
+	// Memory and CpuShares should never be negative
+	if svc.MemoryLimit < 0 {
+		cfg.Memory = 0
+	} else {
+		cfg.Memory = int64(svc.MemoryLimit)
+	}
+
+	if svc.CPUShares < 0 {
+		cfg.CPUShares = 0
+	} else {
+		cfg.CPUShares = svc.CPUShares
+	}
+
+	hcfg.LogConfig.Type = a.dockerLogDriver
+	hcfg.LogConfig.Config = a.dockerLogConfig
+
+	// CC-1848: set core ulimit to 0
+	hcfg.Ulimits = []dockerclient.ULimit{
+		{
+			Name: "core",
+			Soft: 0,
+			Hard: 0,
+		},
+	}
+	return cfg, hcfg, nil
 }
